[Bench] Rename reference .TU.ref files .TU.$HOST_$GPUARCH

+++ tests/GPU/DomainFlowLES/check_perf.sh
@@ -2,7 +2,7 @@
 check()
 {
    [ "$2" = -nsys ] && exit
-   if [ ! -f $1.TU ]
+   if [ ! -f $1.TU ] || [ "`grep 'Arret des processes' $jdd.out_err`" = "" ]
    then
       echo "==================================================="
       echo "Performance is KO for $1 on $2 : case does not run!"
@@ -10,25 +10,32 @@ check()
       exit -1
    fi
    TU=$1.TU
+   TU_REF=$1.TU.$2
+   if [ ! -f $TU_REF ]
+   then
+      mv -f $TU $TU_REF && [ "$TRUST_SCM" = 1 ] && git add $TU_REF
+      echo "Creating new reference $TU_REF"
+      exit
+   fi   
    # Non regression faiblement testee sur le nombre d'iterations des solveurs
-   ref=`awk '/Iterations/ && /solveur/ {print $NF}' $1.TU.ref_$2`
+   ref=`awk '/Iterations/ && /solveur/ {print $NF}' $TU_REF`
    new=`awk '/Iterations/ && /solveur/ {print $NF}' $TU`
    [ $ref != $new ] && echo "Solver convergence is different ($ref != $new) ! Possible regression..." && exit -1
 
-   ref=`awk '/Secondes/ && /pas de temps/ {print $NF}' $1.TU.ref_$2`
+   ref=`awk '/Secondes/ && /pas de temps/ {print $NF}' $TU_REF`
    new=`awk '/Secondes/ && /pas de temps/ {print $NF}' $TU`
    echo $ref $new | awk '// {if (2*($2-$1)/($1+$2)>0.05) {exit 1}}' # On verifie qu'on ne depasse pas +5% de la performance
    err=$?
    ecart=`echo $ref $new | awk '// {printf("%2.1f\n",200*($2-$1)/($1+$2))}'`
    if [ $err = 1 ]
    then
-      sdiff $1.TU.ref_$2 $TU
+      sdiff $TU_REF $TU
       echo "=========================================="
       echo "Performance is KO ($ecart%) for $1 on $2 !"
       echo "=========================================="
    else
       echo "Performance is OK ($ecart%) $new s < $ref s (reference) for $1 on $2"
-      [ `echo "$ecart<-0.99" | bc -l` = 1 ] && echo "Performance is improved so $1.TU.ref_$2 is updated !" && cp $TU $1.TU.ref_$2
+      [ `echo "$ecart<-0.99" | bc -l` = 1 ] && echo "Performance is improved so $TU_REF is updated !" && cp $TU $TU_REF
    fi
 }
 run()
@@ -60,9 +67,14 @@ do
 done
 
 # Liste des machines:
-[ "$1" = -nsys ] && run -nsys
-[ $HOST = is157091 ]       && run a6000
-[ $HOST = topaze ]         && run a100
-[ $HOST = irene-amd-ccrt ] && run v100
+if [ "$1" = -nsys ]
+then
+   run -nsys
+else
+   GPU_ARCH=""
+   [ "$TRUST_USE_CUDA" = 1 ] && GPU_ARCH=_cc$TRUST_CUDA_CC
+   [ "$TRUST_USE_ROCM" = 1 ] && GPU_ARCH=_$ROCM_ARCH
+   run $HOST$GPU_ARCH
+fi
 # clean
 rm -f *.sauv *.lml *.sqlite *.nsys-rep
