+++ tests/GPU/OpenMP_Iterateur/check_perf.sh
@@ -10,20 +10,27 @@ check()
       exit -1
    fi
    TU=$1.TU
-   ref=`awk '/Secondes/ && /pas de temps/ {print $NF}' $1.TU.ref_$2`
+   TU_REF=$1.TU.$2
+   if [ ! -f $TU_REF ]
+   then
+      mv -f $TU $TU_REF && [ "$TRUST_SCM" = 1 ] && git add $TU_REF
+      echo "Creating new reference $TU_REF"
+      exit
+   fi 
+   ref=`awk '/Secondes/ && /pas de temps/ {print $NF}' $TU_REF`
    new=`awk '/Secondes/ && /pas de temps/ {print $NF}' $TU`
    echo $ref $new | awk '// {if (($2-$1)/($1+$2)>0.05) {exit 1}}' # On verifie qu'on ne depasse pas +5% de la performance
    err=$?
    if [ $err = 1 ]
    then
-      sdiff $1.TU.ref_$2 $TU
+      sdiff $TU_REF $TU
       echo "================================"
       echo "Performance is KO for $1 on $2 !"
       echo "================================"
    else
       ecart=`echo $ref $new | awk '// {printf("%2.1f\n",-2*($1-$2)/($1+$2)*100)}'`
       echo "Performance is OK ($ecart%) $new s < $ref s (reference) for $1 on $2"
-      [ `echo "$ecart<-0.5" | bc -l` = 1 ] && echo "Performance is improved so $1.TU.ref_$2 is updated !" && cp $TU $1.TU.ref_$2
+      [ `echo "$ecart<-0.5" | bc -l` = 1 ] && echo "Performance is improved so $TU_REF is updated !" && cp $TU $TU_REF
    fi
 }
 run()
@@ -41,16 +48,22 @@ run()
       make_PAR.data $jdd $np 1>/dev/null 2>&1
       rm -f PAR_$jdd.TU
       trust $nsys PAR_$jdd $np 1>$PAR_jdd.out_err 2>&1
-      check PAR_$jdd $gpu
+      check PAR_$jdd $gpu"x$np"
    fi
 }
 # Liste des machines:
-[ "$1" = -nsys ] && run -nsys
-[ $HOST = is157091 ] && run a6000 && run 1xa6000 2
-[ "`hostname`" = petra ] && run a5000 && run 2xa5000 2
-[ $HOST = is246827 ] && run a3000 && echo "Verifier que le portage se charge, debrancher et rebrancher la prise pour que le GPU tourne a fond !"
-[ $HOST = topaze ]   && run a100   && run 4xa100 4   && run 8xa100 8 OpenMP_Iterateur_BENCH_AmgX_10
-[ $HOST = adastra ]  && run MI250X && run 4xMI250X 4 && run 8xMI250X 8 OpenMP_Iterateur_BENCH_rocALUTION_10
-[ $HOST = jean-zay ] && run v100-32g && run 4xv100-32g 4
-[ $HOST = irene-amd-ccrt ] && run v100
-[ $HOST = is247056 ] && run a5000_is247056 && run 2xa5000_is247056 2
+if [ "$1" = -nsys ]
+then
+   run -nsys
+else
+   [ "$TRUST_USE_CUDA" = 1 ] && GPU_ARCH=_cc$TRUST_CUDA_CC
+   [ "$TRUST_USE_ROCM" = 1 ] && GPU_ARCH=_$ROCM_ARCH
+   run $HOST$GPU_ARCH
+   # Multi-gpu:
+   [ $HOST = is157091 ]     && run $HOST$GPU_ARCH 2
+   [ "`hostname`" = petra ] && run $HOST$GPU_ARCH 2
+   [ $HOST = topaze ]       && run $HOST$GPU_ARCH 4 && run $HOST$GPU_ARCH 8 OpenMP_Iterateur_BENCH_AmgX_10
+   [ $HOST = adastra ]      && run $HOST$GPU_ARCH 4 && run $HOST$GPU_ARCH 8 OpenMP_Iterateur_BENCH_rocALUTION_10
+   [ $HOST = jean-zay ]     && run $HOST$GPU_ARCH 4
+   [ $HOST = is247056 ]     && run $HOST$GPU_ARCH 2
+fi
